+++ example/lib/main.dart
@@ -41,9 +41,8 @@ class _ExampleNavigationScreenState extends State<ExampleNavigationScreen> {
                 label: "Tab 1",
                 icon: Icons.activity,
                 selectedIcon: Icons.activity_fill,
-                builder: (context, footer) => SliverScreen(
+                builder: (context) => SliverScreen(
                     gutter: false,
-                    footer: footer,
                     header: Bar(titleText: "Tab 1"),
                     sliver: SListView.builder(
                         childCount: 10,
@@ -56,9 +55,8 @@ class _ExampleNavigationScreenState extends State<ExampleNavigationScreen> {
                 label: "Tab 2",
                 icon: Icons.address_book,
                 selectedIcon: Icons.address_book_fill,
-                builder: (context, footer) => SliverScreen(
+                builder: (context) => SliverScreen(
                     gutter: false,
-                    footer: footer,
                     header: Bar(titleText: "Tab 2"),
                     sliver: SGridView.builder(
                         crossAxisCount: 3,
@@ -74,8 +72,7 @@ class _ExampleNavigationScreenState extends State<ExampleNavigationScreen> {
                 label: "Tab 3",
                 icon: Icons.gear_six,
                 selectedIcon: Icons.gear_six_fill,
-                builder: (context, footer) => FillScreen(
-                    footer: footer,
+                builder: (context) => FillScreen(
                     gutter: false,
                     header: Bar(titleText: "Tab 3"),
                     child: Center(